refactor(oas): extract OpenAPI transformer
